Retry if getJobs failes [HZG-409] (#4547)
jira -https://hazelcast.atlassian.net/browse/HZG-409

Problem:
getJobs fails if the master node changes during the request.

Solution:
Retry the request on the client side.
Retry the request on the node side.

Changes:
1. Introduces InvokeOnMembers invocation, which executes an operation on
all cluster members and aggregates the results using the provided map
and reduce functions.
2. Updated the invocation strategy for GetJobIdOperation: it now invokes
on the master node for regular jobs and on all nodes for light jobs.

Important notes:
1. The client might not be able to send the request directly to the
master node. For example, the client is using a single-node connection.
That is why the client sends a request to a random node, not directly to
the master.
3. A "node"(in the sense of address) may stop being the master after the
client obtains the master’s address but before the request is processed.
If the master restarts and a new node starts at the same address, the
address remains valid, but the new node is no longer the master.

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases

---------

Co-authored-by: Krzysztof Jamróz <79092062+k-jamroz@users.noreply.github.com>
Co-authored-by: James Holgate <130981049+JamesHazelcast@users.noreply.github.com>
GitOrigin-RevId: ec807ed5a2cb39284931e070eaba896ebd735043