Revert Modifier of `AbstractIMapEvent` Constructor [HZG-537] (#4817)
As problem stated in (close #4751), the change in the constructor causes
a change in serialUUID of the class. That can break the user code if
inheritee class already serialized, and we can't break in a minor
release. The PR reverts the changes to prevent any breaking change on
user code.

Checklist:

- [X] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [ ] Label `All Languages Should Check` set if making changes to the
Java client API and it may effect non-Java clients. Setting this label
will auto generate associated issues in all non-Java clients.
- [ ] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [ ] Request reviewers if possible
- [ ] New public APIs have `@Nonnull/@Nullable` annotations
- [ ] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

GitOrigin-RevId: 3f5b74c005598e31752e861c0dbc8b70a8a472de