Relabel `master` from `6.0-SNAPSHOT` to `5.6-SNAPSHOT` [CTT-504] (#4852)
Renames the version on `master` from `6.0-SNAPSHOT` to `5.6-SNAPSHOT`.

This is obviously a large change, so to aid reviewing I've split the PR
into a commit per-topic so that you can review the specific change (e.g.
updating the version in `pom.xml`) once and then just check the rest all
follow the same pattern.

This change also highlighted
[CCT-608](https://hazelcast.atlassian.net/browse/CTT-608) - _for now_
the test is `Ignore`d to allow merge, but logged to be fixed.

Fixes: [CTT-504](https://hazelcast.atlassian.net/browse/CTT-504)

[CTT-504]:
https://hazelcast.atlassian.net/browse/CTT-504?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: da9ed009c81d191ab0124b243c7cea20eb85ccb9