Fix Predicate corruption due to migration during map loading [CTT-614] (#4898)
If a map is configured with a maploader and has any load in progress
during a migration commit operation on the migration source an exception
can be thrown during finalization which can lead to predicate issues.
The affected member will still see itself as owning the migrated
partition wrt queries. Since it no longer owns the data it will return
responses signalling no matches which can override the response from the
actual owner of the partition which may contain matches. The end result
is that entries which should match the predicate are not returned.

To work around this the offending call `recordStore.isEmpty()` has been
replaced with `recordStore.size() == 0`. The `size()` method already
does not have the loading check and so the change minimises behaviour
change for maps.

---

Some investigation has been carried out into the ramifications of
leaving active loads ongoing on a record store even when the member has
removed it from the partition container and cleaned up the contents at
the time of migration commit:

1. Any load started on a record store is not bound to this specific
record store, i.e. after the load completes it will not try to insert
the data into this store. Instead it will create new partition
operations `RemoveFromLoadAllOperation`, `PutFromLoadAllOperation` which
run on the partition thread and so will not have access to the
recordStore they started from since the `MigrationCommitOperation`
already removed the reference. This means we will not encounter a true
memory leak.
2. It is possible (but unlikely) that a partition p is migrated from a
member but the background load operations puts data into a new
recordStore in the partition container for p on this old owner. Let us
say we are migrating partition p from member A to B when there is
background loading on A, then the sequence of events could look like:
* Data transfer completes from `A` to `B`, partition tables on all
members all still say `A` owns `p`
* Commit sent to B, it updates its partition table, master gets success
response and updates its partition table
    * Update broadcast to other members after some small delay
* `PartitionInvocation` `I` created on B from the background loading
targeting `p` , it is invoked and target initialized to local member but
not placed on op queue
    * `B` updates partition table
* `B` schedules commit finalization op as it is the source of the
migration
    * The operation wrapped by `I` gets added to the same queue

The consequence of the sequence in 2 occurring is a new recordStore
being created on A and some data being inserted into it even though A
should no longer have a replica of p. This data would be overwritten in
any subsequent migration back to A and shouldn't pose any issues except
a small amount of wasted memory. In the vast majority of cases we'd
expect the invocations created by the background loading tasks to be
redirected to the new owner once the partition table on A is updated.

This may cause problems for the initial load as we would potentially
have 2 initial loads happening concurrently on 2 separate members, if
the initial load on B completed before A and some data was removed by a
client then the other initial load on A may insert it back. Fresh data
from a client could possibly be overwritten too. During the initial load
`BasicRecordStoreLoader#loadValues` is called in batches with
`replaceExistingValues == false`, with this flag as false any keys which
already exist in the map are ignored, but we could have the sequence
* Key K is loaded on old owner A, it isn't currently in the migrated
partition p as B hasn't loaded it yet so A considers it fine to add
* K is loaded and inserted on B
* Initial load on B finishes
* Client writes new data for key K
* A overwrites new data with stale data

Although again this should be unlikely. I think it is also the existing
behaviour and is unaffected by this change.

---

Fixes https://hazelcast.atlassian.net/browse/CTT-614

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Request reviewers if possible

---------

Co-authored-by: Krzysztof Jamróz <krzysztof.jamroz@hazelcast.com>
GitOrigin-RevId: 0ecac342c3aeb039aac8c56c6f231fb4dd815459