Relax near cache invalidation test to allow accepted ordering race condition [CTT-461] (#4862)
InvalidationEventBalancingTest was recently introduced as part of
https://github.com/hazelcast/hazelcast-mono/pull/4432 and has been
failing fairly regularly, see
https://github.com/hazelcast/hazelcast/issues/26516. It has caught a pre
existing race condition on invalidations which causes them to arrive out
of order to listeners. This then affects the miss count in
RepairingHandler which expects sequence orders to be incrementing.

In BatchInvalidator each data structure has an invalidation queue which
contains invalidations for all partitions owned by the member. It is
currently possible for the following sequence to happen:

1. Thread A locks queue, polls invalidations taking one for partition p
and releases
2. Thread C adds new invalidation for partition p
3. Thread B locks queue, polls invalidations (including newly added one
for p) and releases
4. Thread B sends events
5. Thread A sends events

The result of this is invalidations for the same partition arriving out
of order. This ordering relaxation is a trade off with queue drain
latency and it was decided the bad ordering was acceptable.

There was also an issue with the test itself occasionally failures due
to timeouts. This was because it was presumed the batch size would
always be the configured value whereas the batch size can actually vary.
So this has also been fixed to instead count the number of individual
invalidations instead of the batches.

Fixes https://hazelcast.atlassian.net/browse/CTT-461

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Request reviewers if possible
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

GitOrigin-RevId: 212e378bbf0f6a336ca98490546fd9740eb992b4