Replace `StringUtils#trim` with `String#strip` _where possible_ (#4916)
In https://github.com/hazelcast/hazelcast/pull/25804,
`StringUtils#trim`'s implementation was updated to make it a `null`-safe
wrapper around `String#strip`.

In _most_ usages we don't actually _need_ the `null`-safety and can
instead simply call `String#strip` directly.

Changes:
- `StringUtils#trim` -> `String#strip` _where possible_
- Added `Nullable` / `Nonnull` annotations to help document where it is
/ isn't required

The rationale for this change is that the method is poorly
named/documented, to be addressed separately in
https://github.com/hazelcast/hazelcast-mono/pull/4915

GitOrigin-RevId: cfed5891c2c94a20db308da0c61d9078771ec5fc