Improve `Queue` `empty-queue-ttl` property description (#4991)
[The
documentation](https://docs.hazelcast.com/hazelcast/latest/data-structures/queue#configuring-queue)
says:
> Used to purge unused or empty queues {...} your queue will be
destroyed if it stays empty or unused for that time.

The scheduled eviction task was _only_ triggered when elements were
removed, not when the `Queue` was first constructed - so unused `Queue`s
would not be removed.

Note - there was [a discussion on changing the
implementation](https://github.com/hazelcast/hazelcast-mono/pull/4943)
but as a change in behaviour / breaking change, safer to document the
_actual_ behaviour.

Corresponding [docs PR](https://github.com/hazelcast/hz-docs/pull/1863).

Fixes: https://github.com/hazelcast/hazelcast/issues/26525,
https://github.com/hazelcast/imdg-docs/issues/310
Closes: https://github.com/hazelcast/hazelcast-mono/pull/4943
GitOrigin-RevId: e95dc972e9bded6709114cd3b8621e390c879692