Fix race condition in StreamKafkaPTest [CTT-620] (#4984)
StreamKafkaPTest#when_processingGuaranteeAtLeastOnceAndJobResumedAfterSuspension_then_readFromPartitionsInitialOffsets
has been flaky because of a race condition on the insertion of the
initial records.

The test triggers the insertion asynchronously and moves on, if the
offset seek occurs before these initial records are inserted Kafka will
treat it as an invalid offset with behaviour defined by the
[auto.offset.reset](https://kafka.apache.org/documentation/#consumerconfigs_auto.offset.reset)
consumer config. The default behaviour is to just reset to the latest
offset. The producer batches records and sends unfilled batches after a
linger which configurable with consumer property
[linger.ms](https://kafka.apache.org/documentation/#producerconfigs_linger.ms)
so a delay between the producer send call and the actual record batch
being sent can be introduced. Increasing this linger value causes the
test to fail every time.

The fix is just to wait for the ack on the initial record population.

Fixes https://hazelcast.atlassian.net/browse/CTT-620,
https://github.com/hazelcast/hazelcast/issues/26341

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Request reviewers if possible
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

GitOrigin-RevId: 7d200eab64058cc57e1fc9a1333951add83f2d5b