JobOperationFaultBounceTest.clientGetAllJobs flaky test [CTT-660] (#5072)
jira - https://hazelcast.atlassian.net/browse/CTT-660

GetJobIdsOperation may run on a node that is shutting down. During
shutdown, ClusterService.masterAddress is set to null. At that moment,
the operation behaves as if the node is not the master, which causes the
test to fail.

JobCoordinationService.getAllJobIds is only supposed to run on the
master. This is already checked in GetJobIdsOperation via
isRequiredMasterExecution.
During shutdown the current node is still technically the master, even
though masterAddress becomes null. So let's allow
JobCoordinationService.getAllJobIds to execute as if the node is still
the master even it is in the process of shutting down.

Compatibility with previous versions:
JobCoordinationService.getAllJobIds is a new method. Older version nodes
do not call it.

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `All Languages Should Check` set if making changes to the
Java client API and it may effect non-Java clients. Setting this label
will auto generate associated issues in all non-Java clients.
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

GitOrigin-RevId: 12e9296bb6298f93fd8e944766edf38a2f3cba38