Retry on AttributeNotFoundException in MetricTestUtils [CTT-430] (#5136)
PartitionsIndexedTest occasionally fails with uncaught
`AttributeNotFoundException`. There is no guarantee that all the
attributes for an MBean are published atomically, it seems that it is
possible for the MBean to exist but the attribute to be unpublished.
This change means the test retries on `AttributeNotFoundException` in
addition to the existing behaviour of retrying on
`InstanceNotFoundException`.

Fixes:
 - https://hazelcast.atlassian.net/browse/CTT-430
 - https://github.com/hazelcast/hazelcast-mono/issues/4466

Checklist:
- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Request reviewers if possible
GitOrigin-RevId: 091bde451d377048369c504254dfccba31e21eec