Explicit throw if ExplicitSuspicionOp received without master set [5.6.0] (#5214)
Backport of https://github.com/hazelcast/hazelcast-mono/pull/5213

Currently if this happens the operation fails with a NPE. This change
checks for this situation and throws an explicit IllegalStateException
with a better message if encountered. This situation occurred in 2
separate nightly pipeline runs but has not been reproduced locally. This
change will not affect any platform behaviour except for the logging
output as the operation will still fail with an unhandled runtime
exception. A new ticket will be created to investigate further why this
happens and to add null handling in the `MembersViewMetadata#equals`
method.

Fixes:
 - https://hazelcast.atlassian.net/browse/CTT-673
 - https://github.com/hazelcast/hazelcast-mono/issues/5126
 - https://hazelcast.atlassian.net/browse/CTT-672
 - https://github.com/hazelcast/hazelcast/issues/26530

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Request reviewers if possible

Co-authored-by: Thomas Ball <thomas.ball@hazelcast.com>
GitOrigin-RevId: 6bde57aa5f339f9b911f5a3db8e4079bde829601