Bound the time spent for on heap near cache doExpiration [HZG-185] (#4683)
Currently any near cache which uses an on heap in memory format has a
performance defect when a ttl or max idle is configured. The existing
behaviour is to loop over every entry in the cache checking for
expiration every 5 seconds. With a long ttl set (e.g. 1 day) on a cache
with many entries there is a large amount of wasted cpu time
continuously traversing the cache. This PR introduces a time bound on
how much cpu is spent per expiration cycle on a near cache. If there are
too many elements in the cache to traverse at once the next cycle will
resume from the last examined record.

Regardless of the number of expired entries in the cache we will never
return an expired record (this is existing behaviour) so having this
maximum is safe.

Fixes https://hazelcast.atlassian.net/browse/HZG-185
See discussion at
https://hazelcast.atlassian.net/browse/HZG-185?focusedCommentId=112528

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Request reviewers if possible

---------

Co-authored-by: Ahmet Mircik <ahmetmircik@gmail.com>
GitOrigin-RevId: 28c703354ef9a73fcfbd96f4a6c77758f0d4356c