Fix false positive `CVE-2017-15945` dependency suppression (#5563)
In https://github.com/hazelcast/hazelcast-mono/pull/4524 the regex was
updated to suppress the (false positive) in the EE extension, but the
regex was subtly wrong - the modules are named:
- `hazelcast-jet-cdc-mysql`
- `hazelcast-enterprise-cdc-sql`

Not the same with an extra `enterprise` inserted in the EE variant as
the regex expected.

For testing, [re-ran the
job](https://github.com/hazelcast/hazelcast-mono/actions/runs/19281432431)
- notice how `CVE-2017-15945` no longer features vs [the previous
`master`
execution](https://github.com/hazelcast/hazelcast-mono/actions/runs/19200478324).

GitOrigin-RevId: 55c349e6bcf6199cf592917ca0c5eeebc0a5b7d3