Don't shade Java-11 specific code from dependencies (#5558)
Some dependencies (e.g. Jackson) offer multi-release JARs -
specifically, they bundle different versions of the same class for use
on Java `11`, `17` & `21`.

We shade Jackson, which means in `META-INF\versions` of the Hazelcast
JAR we have those classes.

However - as _we_ require a minimum of Java `17`, we don't need to
include the Java `11` multi-release components as they will never be
used.

GitOrigin-RevId: 35c50d8aa433385c0f5e56b94a33e00b202a8dca