Always use latest Sonar plugin version (#5656)
This PR updates our Sonar configuration to always use the latest plugin
version instead of hardcoding it in `pom.xml`. Since we use SonarCloud
and automatic updates are only applied on the `master` branch, this
approach makes it easier to keep the analysis configuration in sync with
the cloud version.

GitOrigin-RevId: 6fb9d613c067221d347beecb69635f487e82b5f1