Support Job namespace for entry processors submitted by jet jobs [CTT-817] (#5502)
Currently we are encountering serialization errors for Jet jobs with an
IMap EntryProcessor sink using custom classpath resources even if the
resources are attached to the job correctly either via a UCN or
directly. The affected factory methods in scope for this change are
`Sinks#mapWithMerging/mapWithUpdate/mapWithEntryProcessor`. This change
only handles the case of using UCN to share the required classpath
resources.

This change makes the compromise that if security is enabled and the EP
is in the job UCN then the UCN must have global use permissions. Otherwise
results are undefined, the job may fail or it may run with map backups failing.
This will be addressed in future in CTT-844.

The root cause is the created EntryProcessors are not aware of the job
namespace if it has been set and they default to using the associated
IMap UCN (if one exists, the default otherwise). This change means that
if the job namespace has been set then it is tried first when
deserializing the relevant fields which may use custom code. If the
classes are not found under this namespace we fallback to the existing
behaviour and use the IMap UCN.

The changes here are derived from the PoC in
https://github.com/hazelcast/hazelcast-mono/pull/5437

Fixes: https://hazelcast.atlassian.net/browse/CTT-817

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Label `All Languages Should Check` set if making changes to the
Java client API and it may effect non-Java clients. Setting this label
will auto generate associated issues in all non-Java clients.
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases

---------

Co-authored-by: Krzysztof Jamróz <krzysztof.jamroz@hazelcast.com>
GitOrigin-RevId: fca3680e7f16af5b28439c7cfdfea682cc44c2e6