Upgrade `org.lz4:lz4-java` to address `CVE-2025-12183` (#5691)
Addresses
[CVE-2025-12183](https://nvd.nist.gov/vuln/detail/CVE-2025-12183).

Of note - this wasn't picked by Dependabot because it [doesn't support
the relocated nature of the updated
artefact](https://github.com/dependabot/dependabot-core/issues/1947).

We can't use the relocated group directly because we only include to
manage the version of a transitive dependency - an alternative would be
to exclude the `org.lz4` variant and manually declare the `at.yawk.lz4`
variant, which would be much more work.

Post-merge actions:
- [ ] backport

GitOrigin-RevId: 96c8d7e8617f3976e53632299695045ada45c3b9